+++ .gitignore
@@ -5,3 +5,5 @@
 /.vscode
 /source/__pycache__
 /pyqt
+/source/py_ui/__pycache__
+/source/tools/__pycache__
