+++ source/tools/empty_field.py
